Add files via upload
These are the files I created in eclipse, I am not sure if they are fully correct or not but I did the best I could based off my knowledge of code and piecing together codes in the current project in order to create this. Please feel free to edit this and make it better. It is pretty basic and some errors exist but the html's pull up on the eclipse web browser and not through localhost. I need to figure out how to connect the shop button on the homepage to go to the products page in the project.